add code to verify saved receipts
review access will also be reevaluated.
currently not aware of consumables and untested for subscriptions.

this also changes the behavior of the completion handler for SSKProducts:
now if a verification failed or the review request was denied the
completion handler is called with NO as parameter instead of calling
the error handler with nil as error.