some theoretical code to allow purchases using codes
the basic idea here is to be able to redeem coupons just like the
app store already allows for apps, but given that StoreKit does not
support such a thing it is loaded off to a developer-controlled server.

this commit might also fix a very simple exploit in the previous json
handling code as I think it was possible to pass the check with
an empty document.

also changes the "error" bahavior again for product callbacks: if there
is a exception embedded in the json and the status indicates an error,
the error handler is called back with a custom error containing the
string from exception in the json as NSLocalizedDescriptionKey.
I know that the string probably is not localized but this is the key
used in most of the other code so for now this is ok.