add function to redeem code using its productIdentifier
if the list of available IAPs was not yet received (or you want to
use this code without StoreKit for whatever reason) you can use
this convenience method. We only need the identifier and any product,
so if we don't have a product just create a new one as temporary
worker object.
also fixes custom errors to use NSLocalizedFailureReasonErrorKey
instead of NSLocalizedDescriptionKey for convenience reasons, fixes
some useless error message in my application.