allow to use SSKeychain instead of SFHFKeychainUtils
currently developers choice, but since I don't want to ship two helper
classes and ShareKit has switched to SSKeychain (without any notice,
though I saw it coming in the respective pull request) I wanted to
let the developer pick the one. SFHFKeychainUtils are quite popular
and used by some third party libraries, but as the only one I use
made the switch no need to me to pull it back in.

not much changes, if you want to use SFHFKeychainUtils define
USE_SFHFKEYCHAIN and if you want to use SSKeychain define USE_SSKEYCHAIN.
if none is defined the latter one is used, if both are defined
SFHFKeychainUtils is used. a prefix header should be used to make sure
the define is set before the fallback is chosen. especially if you cannot
modify SSKConfig.h because you use SSKTk as git submodule.